Clean code, bug fixes, refactoring and updated readme for raspberrypi and real thymio (#4)
Thymio and interface have been combined so the repository only has to be built once. Before this the newinterface had to be built separately. Now it builds automatically when -DARGOS_BUILD_FOR=thymio is set.

Removed unnecessary dependencies.

Added clearer instructions to the readme.

Refactored file and directory names to be all lower case to follow the ARGoS naming convention.